move struct _fluid_mod_t to private header
