+++ src/synth/fluid_mod.h
@@ -24,17 +24,30 @@
 #include "fluidsynth_priv.h"
 #include "fluid_conv.h"
 
-void fluid_mod_clone(fluid_mod_t* mod, fluid_mod_t* src);
+/**
+ * Modulator structure.  See SoundFont 2.04 PDF section 8.2.
+ */
+struct _fluid_mod_t
+{
+  unsigned char dest;           /**< Destination generator to control */
+  unsigned char src1;           /**< Source controller 1 */
+  unsigned char flags1;         /**< Source controller 1 flags */
+  unsigned char src2;           /**< Source controller 2 */
+  unsigned char flags2;         /**< Source controller 2 flags */
+  double amount;                /**< Multiplier amount */
+  /* The 'next' field allows to link modulators into a list.  It is
+   * not used in fluid_voice.c, there each voice allocates memory for a
+   * fixed number of modulators.  Since there may be a huge number of
+   * different zones, this is more efficient.
+   */
+  fluid_mod_t * next;
+};
+
 fluid_real_t fluid_mod_get_value(fluid_mod_t* mod, fluid_channel_t* chan, fluid_voice_t* voice);
+
+#ifdef DEBUG
 void fluid_dump_modulator(fluid_mod_t * mod);
-
-#define fluid_mod_has_source(mod,cc,ctrl)  \
-( ((((mod)->src1 == ctrl) && (((mod)->flags1 & FLUID_MOD_CC) != 0) && (cc != 0)) \
-   || ((((mod)->src1 == ctrl) && (((mod)->flags1 & FLUID_MOD_CC) == 0) && (cc == 0)))) \
-|| ((((mod)->src2 == ctrl) && (((mod)->flags2 & FLUID_MOD_CC) != 0) && (cc != 0)) \
-    || ((((mod)->src2 == ctrl) && (((mod)->flags2 & FLUID_MOD_CC) == 0) && (cc == 0)))))
-
-#define fluid_mod_has_dest(mod,gen)  ((mod)->dest == gen)
+#endif
 
 
 #endif /* _FLUID_MOD_H */
